+++ CHANGELOG.md
@@ -1,5 +1,10 @@
 # Piral Changelog
 
+## 1.12.4 (tbd)
+
+- Updated dependencies
+- Added `piral-ng/extend-rspack` helper for explicit rspack support in `piral-ng`
+
 ## 1.12.3 (August 18, 2026)
 
 - Fixed using development condition in production dependency resolution (#868)